+++ src/main/java/xt9/inworldcrafting/common/crafttweaker/FireCrafting.java
@@ -3,14 +3,10 @@ package xt9.inworldcrafting.common.crafttweaker;
 import crafttweaker.api.item.IIngredient;
 import crafttweaker.api.item.IItemStack;
 import crafttweaker.api.minecraft.CraftTweakerMC;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.NonNullList;
 import stanhebben.zenscript.annotations.ZenClass;
 import stanhebben.zenscript.annotations.ZenMethod;
-import xt9.inworldcrafting.ModConfig;
-import xt9.inworldcrafting.common.event.EntityReplacer;
+import xt9.inworldcrafting.common.event.EntityMatcher;
 import xt9.inworldcrafting.common.recipe.BurnItemRecipe;
-import xt9.inworldcrafting.common.util.ModClashInfo;
 
 /**
  * Created by xt9 on 2019-01-19.
@@ -28,14 +24,7 @@ public class FireCrafting {
         /* Inputs should only be items or oredicts */
         if(ingredient.getLiquids().size() > 0) { return; }
 
-        ModClashInfo info = new ModClashInfo();
-        if(!ModConfig.blacklistDisabled) {
-            info.checkForClashes(ingredient);
-        }
-
-        if(!info.isModClashed()) {
-            EntityReplacer.allValidInputs.add(ingredient);
-            BurnItemRecipe.addRecipe(CraftTweakerMC.getItemStack(output), ingredient, ingredient.getAmount(), ticks);
-        }
+        EntityMatcher.allValidInputs.add(ingredient);
+        BurnItemRecipe.addRecipe(CraftTweakerMC.getItemStack(output), ingredient, ingredient.getAmount(), ticks);
     }
 }