+++ changelog.md
@@ -1,5 +1,10 @@
 # 1.12.2
 
+## 1.0
+- Removed the item blacklist, changed approach to crafting so we don't cannibalized custom EntityItem classes. (Might be a bit more expensive on the cpu, so marking this version as a beta)
+- Fixed a bug where Fluid to Item recipes that consumes the fluid could craft in flowing fluid.
+- Inputs that are not consumed during crafting will now properly update it's shrunken state.
+
 ## 0.13
 - Added NBT matching for ingredients